feat: add header navigation in the app
