+++ src/ts/components/header/header-unauth/header-unauth.ts
@@ -18,7 +18,7 @@ export class HeaderUnauth extends Control {
     });
   }
 
-  changeMainPages(page: string): void {
+  private changeMainPages(page: string): void {
     state.setNewPage(page);
   }
 }
