+++ src/ts/components/header/header.scss
@@ -4,6 +4,13 @@
   justify-content: space-between;
   align-items: center;
   padding: 1rem;
+  &-img {
+    cursor: pointer;
+    transition: all 0.5s;
+    &:hover {
+      transform: scale(1.1);
+    }
+  }
 }
 
 .img-thumbnail {
