+++ src/ts/components/header/nav-menu/nav-menu.ts
@@ -1,5 +1,5 @@
 import Control from '../../../common/control';
-import { PagesList, state } from '../../../common/state';
+import { state } from '../../../common/state';
 import './nav-menu.scss';
 
 export class NavBar extends Control {
@@ -18,7 +18,7 @@ export class NavBar extends Control {
     });
   }
 
-  changeMainPages(page: string): void {
+  private changeMainPages(page: string): void {
     state.setNewPage(page);
   }
 }
